Move `ErrType` and related utils into its own module.
