+++ src/lib/Err.hs
@@ -1,0 +1,128 @@
+-- Copyright 2021 Google LLC
+--
+-- Use of this source code is governed by a BSD-style
+-- license that can be found in the LICENSE file or at
+-- https://developers.google.com/open-source/licenses/bsd
+
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE FlexibleContexts #-}
+
+module Err (Err (..), ErrType (..), Except, SrcPos, SrcCtx,
+            throw, throwIf, modifyErr,
+            addContext, addSrcContext, catchIOExcept, liftEitherIO,
+            assertEq, ignoreExcept, pprint, docAsStr) where
+
+import Control.Exception hiding (throw)
+import Control.Monad
+import Control.Monad.Except hiding (Except)
+import Data.Text (unpack)
+import Data.Text.Prettyprint.Doc.Render.Text
+import Data.Text.Prettyprint.Doc
+import GHC.Stack
+import System.Environment
+import System.IO.Unsafe
+
+data Err = Err ErrType SrcCtx String  deriving (Show, Eq)
+data ErrType = NoErr
+             | ParseErr
+             | TypeErr
+             | KindErr
+             | LinErr
+             | UnboundVarErr
+             | RepeatedVarErr
+             | CompilerErr
+             | IRVariantErr
+             | NotImplementedErr
+             | DataIOErr
+             | MiscErr
+             | RuntimeErr
+               deriving (Show, Eq)
+
+type Except = Either Err
+
+type SrcPos = (Int, Int)
+type SrcCtx = Maybe SrcPos
+
+throw :: MonadError Err m => ErrType -> String -> m a
+throw e s = throwError $ Err e Nothing s
+
+throwIf :: MonadError Err m => Bool -> ErrType -> String -> m ()
+throwIf True  e s = throw e s
+throwIf False _ _ = return ()
+
+modifyErr :: MonadError e m => m a -> (e -> e) -> m a
+modifyErr m f = catchError m \e -> throwError (f e)
+
+addContext :: MonadError Err m => String -> m a -> m a
+addContext s m = modifyErr m \(Err e p s') -> Err e p (s' ++ "\n" ++ s)
+
+addSrcContext :: MonadError Err m => SrcCtx -> m a -> m a
+addSrcContext ctx m = modifyErr m updateErr
+  where
+    updateErr :: Err -> Err
+    updateErr (Err e ctx' s) = case ctx' of Nothing -> Err e ctx  s
+                                            Just _  -> Err e ctx' s
+
+catchIOExcept :: (MonadIO m , MonadError Err m) => IO a -> m a
+catchIOExcept m = (liftIO >=> liftEither) $ (liftM Right m) `catches`
+  [ Handler \(e::Err)           -> return $ Left e
+  , Handler \(e::IOError)       -> return $ Left $ Err DataIOErr   Nothing $ show e
+  , Handler \(e::SomeException) -> return $ Left $ Err CompilerErr Nothing $ show e
+  ]
+
+liftEitherIO :: (Exception e, MonadIO m) => Either e a -> m a
+liftEitherIO (Left err) = liftIO $ throwIO err
+liftEitherIO (Right x ) = return x
+
+ignoreExcept :: HasCallStack => Except a -> a
+ignoreExcept (Left e) = error $ pprint e
+ignoreExcept (Right x) = x
+
+assertEq :: (MonadError Err m, Show a, Pretty a, Eq a) => a -> a -> String -> m ()
+assertEq x y s = if x == y then return ()
+                           else throw CompilerErr msg
+  where msg = "assertion failure (" ++ s ++ "):\n"
+              ++ pprint x ++ " != " ++ pprint y ++ "\n"
+
+-- === small pretty-printing utils ===
+-- These are here instead of in PPrint.hs for import dependency reasons
+
+pprint :: Pretty a => a -> String
+pprint x = docAsStr $ pretty x
+
+docAsStr :: Doc ann -> String
+docAsStr doc = unpack $ renderStrict $ layoutPretty layout $ doc
+
+layout :: LayoutOptions
+layout = if unbounded then LayoutOptions Unbounded else defaultLayoutOptions
+  where unbounded = unsafePerformIO $ (Just "1"==) <$> lookupEnv "DEX_PPRINT_UNBOUNDED"
+
+-- === instances ===
+
+instance MonadFail (Either Err) where
+  fail s = Left $ Err CompilerErr Nothing s
+
+instance Exception Err
+
+instance Pretty Err where
+  pretty (Err e _ s) = pretty e <> pretty s
+
+instance Pretty ErrType where
+  pretty e = case e of
+    -- NoErr tags a chunk of output that was promoted into the Err ADT
+    -- by appending Results.
+    NoErr             -> ""
+    ParseErr          -> "Parse error:"
+    TypeErr           -> "Type error:"
+    KindErr           -> "Kind error:"
+    LinErr            -> "Linearity error: "
+    IRVariantErr      -> "Internal IR validation error: "
+    UnboundVarErr     -> "Error: variable not in scope: "
+    RepeatedVarErr    -> "Error: variable already defined: "
+    NotImplementedErr -> "Not implemented:"
+    CompilerErr       ->
+      "Compiler bug!" <> line <>
+      "Please report this at github.com/google-research/dex-lang/issues\n" <> line
+    DataIOErr         -> "IO error: "
+    MiscErr           -> "Error:"
+    RuntimeErr        -> "Runtime error"
