+++ src/lib/Imp.hs
@@ -39,8 +39,9 @@ import qualified Data.Map.Strict as M
 import Builder
 import Syntax
 import Env
+import Err
 import Type
-import PPrint
+import PPrint ()
 import Cat
 import Algebra
 import Util
